+++ src/main/java/dto/status/PlayerStatus.java
@@ -1,3 +1,3 @@
 package dto.status;
 
-public record PlayerStatus(String name, Integer score, Integer bet, boolean isBust, boolean isBlackJack) {}
+public record PlayerStatus(String name, Integer score, Long bet, boolean isBust, boolean isBlackJack) {}
